requirements.txt
numba==0.57.1
numpy==1.24.3
autopep8==2.0.2
pandas==2.1.2
pillow==10.1.0
skorch==0.15.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.65.0
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
ipywidgets==7.7.2
dataframe_image
plotly
timm
lightning